refact(folders): better organized different cloud-providers and core features
