+++ src/core/models/ServerCredentials.ts
