+++ src/core/services/CredentialsService.ts
@@ -1,0 +1,8 @@
+import { ServerCredentials } from '../models/ServerCredentials';
+
+/**
+ * Service responsible for generating all server credentials
+ */
+export interface CredentialsService {
+    generateCredentials(): ServerCredentials;
+}
