+++ src/providers/cloud-providers/aws/ECSServerManagerFactory.ts
@@ -1,0 +1,52 @@
+import { Chance } from "chance";
+import { ServerCommander } from "../../../core/services/ServerCommander";
+import { ConfigManager } from "../../../core/utils/ConfigManager";
+import { PasswordGeneratorService } from "../../../core/services/PasswordGeneratorService";
+import { AWSClientFactory } from "../../services/defaultAWSServiceFactory";
+import { ECSServerManager } from "./ECSServerManager";
+import { DefaultTaskDefinitionService } from "./services/TaskDefinitionService";
+import { DefaultECSServiceManager } from "./services/ECSServiceManager";
+import { DefaultSecurityGroupService } from "./services/SecurityGroupService";
+import { DefaultNetworkService } from "./services/NetworkService";
+import { DefaultEC2InstanceService } from "./services/EC2InstanceService";
+import { DefaultCredentialsService } from "../../services/DefaultCredentialsService";
+import { DefaultEnvironmentBuilderService } from "../../services/DefaultEnvironmentBuilderService";
+import { DefaultTF2ServerReadinessService } from "../../services/DefaultTF2ServerReadinessService";
+
+export interface ECSServerManagerDependencies {
+    configManager: ConfigManager;
+    awsClientFactory: AWSClientFactory;
+    serverCommander: ServerCommander;
+    passwordGeneratorService: PasswordGeneratorService;
+    chance: Chance.Chance;
+}
+
+export class ECSServerManagerFactory {
+    static createServerManager(dependencies: ECSServerManagerDependencies): ECSServerManager {
+        const { configManager, awsClientFactory, serverCommander, passwordGeneratorService, chance } = dependencies;
+        
+        // Create AWS services - they take configManager and awsClientFactory
+        const taskDefinitionService = new DefaultTaskDefinitionService(configManager, awsClientFactory);
+        const ecsServiceManager = new DefaultECSServiceManager(configManager, awsClientFactory);
+        const securityGroupService = new DefaultSecurityGroupService(configManager, awsClientFactory);
+        const networkService = new DefaultNetworkService(configManager, awsClientFactory);
+        const ec2InstanceService = new DefaultEC2InstanceService(configManager, awsClientFactory);
+        
+        // Create core services
+        const credentialsService = new DefaultCredentialsService(passwordGeneratorService, chance);
+        const environmentBuilderService = new DefaultEnvironmentBuilderService();
+        const tf2ServerReadinessService = new DefaultTF2ServerReadinessService(serverCommander);
+        
+        return new ECSServerManager(
+            taskDefinitionService,
+            securityGroupService,
+            ec2InstanceService,
+            ecsServiceManager,
+            networkService,
+            tf2ServerReadinessService,
+            environmentBuilderService,
+            credentialsService,
+            configManager
+        );
+    }
+}
