+++ src/providers/cloud-providers/aws/models/ContainerConfig.ts
