+++ src/providers/cloud-providers/aws/models/NetworkConfiguration.ts
