+++ src/providers/cloud-providers/aws/models/TaskDefinitionConfig.ts
