+++ src/providers/cloud-providers/oracle/OCIServerManager.ts
@@ -1,16 +1,16 @@
-import { logger, tracer, meter } from '../../telemetry/otel';
+import { logger, tracer, meter } from '../../../telemetry/otel';
 import { Span } from '@opentelemetry/api';
 import { containerinstances, core } from "oci-sdk";
-import { getRegionDisplayName, Region, Server, Variant } from "../../core/domain";
-import { ServerStatus } from "../../core/domain/ServerStatus";
-import { ServerAbortManager } from "../../core/services/ServerAbortManager";
-import { ServerCommander } from "../../core/services/ServerCommander";
-import { ServerManager } from "../../core/services/ServerManager";
-import { PasswordGeneratorService } from "../../core/services/PasswordGeneratorService";
-import { ConfigManager } from "../../core/utils/ConfigManager";
-import { waitUntil } from "../utils/waitUntil";
-import { OCICredentialsFactory } from "../../core/services/OCICredentialsFactory";
-import { StatusUpdater } from "../../core/services/StatusUpdater";
+import { getRegionDisplayName, Region, Server, Variant } from "../../../core/domain";
+import { ServerStatus } from "../../../core/domain/ServerStatus";
+import { ServerAbortManager } from "../../../core/services/ServerAbortManager";
+import { ServerCommander } from "../../../core/services/ServerCommander";
+import { ServerManager } from "../../../core/services/ServerManager";
+import { PasswordGeneratorService } from "../../../core/services/PasswordGeneratorService";
+import { ConfigManager } from "../../../core/utils/ConfigManager";
+import { waitUntil } from "../../utils/waitUntil";
+import { OCICredentialsFactory } from "../../../core/services/OCICredentialsFactory";
+import { StatusUpdater } from "../../../core/services/StatusUpdater";
 import { Chance } from "chance";
 
 const chance = new Chance();
