--- src/providers/services/ecs-server-management/services/PasswordGeneratorService.ts
@@ -1,11 +1,0 @@
-/**
- * Service interface for password generation
- */
-export interface PasswordGeneratorService {
-    generatePassword(settings: {
-        alpha?: boolean;
-        length?: number;
-        numeric?: boolean;
-        symbols?: boolean;
-    }): string;
-}
